+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#v0.2.0
+- `current_style_as_preset` method for convenient conversion of a style into a preset
+- New Markdown like table style
+
 #v0.0.7
 - Fix wrong calculation due to bytes count instead of char count
 - Fix panics caused by wrong string splits